requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.2
openai==1.3.9
packaging==23.2
pgvector==0.2.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
django-cors-headers==4.3.1
cachetools==5.3.2
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
pandas==2.0.1
pypdf==4.2.0